+++ src/histograms.cpp
@@ -125,7 +125,7 @@ TGraphErrors nancycs(const int &setnumber){
          {1,0.4,0.3}
      };
 
-     for(int i=0; i<massnumbers.at(setnumber).size(); i++){
+     for(unsigned long i=0; i<massnumbers.at(setnumber).size(); i++){
          temp.SetPoint(temp.GetN(),massnumbers.at(setnumber).at(i),
                                    crosssections.at(setnumber).at(i));
          temp.SetPointError(temp.GetN()-1, 0, crosssectione.at(setnumber).at(i));
